# Performance gains for hex and base64