--- a/package.json
+++ b/package.json
@@ -16,7 +16,7 @@
     "James Halliday <mail@substack.net>"
   ],
   "dependencies": {
-    "base64-js": "^1.3.1",
+    "hextreme": "^1.0.5",
     "ieee754": "^1.2.1"
   },
   "devDependencies": {
@@ -59,7 +59,7 @@
   },
   "scripts": {
     "perf": "browserify --debug perf/bracket-notation.js > perf/bundle.js && open perf/index.html",
-    "perf-node": "node perf/bracket-notation.js && node perf/concat.js && node perf/copy-big.js && node perf/copy.js && node perf/new-big.js && node perf/new.js && node perf/readDoubleBE.js && node perf/readFloatBE.js && node perf/readUInt32LE.js && node perf/slice.js && node perf/writeFloatBE.js && node perf/write-hex.js",
+    "perf-node": "node perf/toString-hex.js && node perf/from-hex.js && node perf/toString-base64.js && node perf/from-base64.js && node perf/bracket-notation.js && node perf/concat.js && node perf/copy-big.js && node perf/copy.js && node perf/new-big.js && node perf/new.js && node perf/readDoubleBE.js && node perf/readFloatBE.js && node perf/readUInt32LE.js && node perf/slice.js && node perf/writeFloatBE.js && node perf/write-hex.js",
     "size": "browserify -r ./ | uglifyjs -c -m | gzip | wc -c",
     "standard": "standard",
     "test": "tape test/*.js test/node/*.js",