@jawj 2024-12-11
This change should also make it very easy to add support for `base64url` (#314) since hextreme supports that.