@jawj 2024-12-12
OK. Kind of disappointing, but I guess I see where you're coming from. I'll start again and probably just do `toString('hex')` for now.